feat:Refactor task hours to use integer type and improve task deletion confirmation

+++ lib/models/task.dart
@@ -2,7 +2,7 @@ import 'package:agrocontrol_app/models/worker.dart';
 
 class Task {
   final DateTime date;
-  final double hours;
+  final int hours;
   final Worker worker;
 
   Task({required this.date, required this.hours, required this.worker});
